# AllAboard_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: AllAboard API
host: 'localhost'
schemes:
  - http
...
paths:
...
  /station/station:
    post:
      tags:
        - station
      summary: Add a new station
      operationId: addStation
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - in: body
          name: station 
          description: Station object 
          required: true
          schema:
            $ref: '#/definitions/Station'
      responses:
        '200':
          description: Station created successfully
  /station/station/{stationId}:
    put:
      tags:
        - station
      summary: Update a station 
      operationId: updateStation
      parameters:
        - name: stationId
          in: path 
          description: ID of the station to update 
          required: true
          type: integer
        - name: station
          in: body
          description: Updated station object 
          required: true
          schema: 
            $ref: '#/definitions/Station'
      responses:
        200:
          description: Station updated successfully
        400: 
          description: Invalid request data
        404:
          description: Station not found
    
    delete: 
      tags:
        - station 
      summary: Delete a station 
      operationId: deleteStation
      parameters:
        - name: stationId
          in: path
          description: ID of the station to delete
          required: true
          type: integer
      responses:
        200:
          description: Station deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Station not found  
...
  /trainClass/trainClass:
    post:
      tags:
        - trainClass
      summary: Add a new train class type
      operationId: addTrainClassType
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - in: body
          name: trainClassType 
          description: Station object 
          required: true
          schema:
            $ref: '#/definitions/TrainClassType'
      responses:
        '200':
          description: Train Class Type created successfully
  /trainClass/trainClass/{classId}:
    put:
      tags:
        - trainClass
      summary: Update a train class type 
      operationId: updateTrainClassType
      parameters:
        - name: classId
          in: path 
          description: ID of the train class type to update 
          required: true
          type: integer
        - name: trainClassType
          in: body
          description: Updated train class type object 
          required: true
          schema: 
            $ref: '#/definitions/TrainClassType'
      responses:
        200:
          description: Train class type updated successfully
        400: 
          description: Invalid request data
        404:
          description: Train class type not found
    
    delete: 
      tags:
        - trainClass 
      summary: Delete a train class type 
      operationId: deleteTrainClassType
      parameters:
        - name: classId
          in: path
          description: ID of the train class type to delete
          required: true
          type: integer
      responses:
        200:
          description: Train class type deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Train class type not found
...
definitions:
...
  Station:
    type: object
    properties:
      station_id:
        type: integer
      station_name:
        type: string
  TrainClassType:
    type: object
    properties:
      train_class_id:
        type: integer
      train_class_name:
        type: string
...